refactor: collapse single-includer adapter concerns into Promises
The four adapter modules only carried class methods for one includer;
future/delay/schedule shared an identical shape. Fold them into
Promises' singleton class and prepend TestingFuture so overrides still
resolve.
